+++ packages/micro-generators/editor-config/lib/bin.js
@@ -1,0 +1,10 @@
+#!/usr/bin/env node
+'use strict';
+
+var core = require('@test-release/core');
+
+var generateEditorConfig = function (task) {
+    core.doSomeCoreFunction(task);
+};
+
+generateEditorConfig('generating from cli');