+++ packages/micro-generators/editor-config/lib/index.d.ts
@@ -1,0 +1,1 @@
+export declare const generateEditorConfig: (task: string) => void;