+++ packages/micro-generators/editor-config/lib/index.js
@@ -1,0 +1,12 @@
+#!/usr/bin/env node
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var core = require('@test-release/core');
+
+var generateEditorConfig = function (task) {
+    core.doSomeCoreFunction(task);
+};
+
+exports.generateEditorConfig = generateEditorConfig;